docs(examples): split cmd/quickstart into Quick Start + e2e variants
README's Quick Start is a 45-line inline snippet; the runnable
program in cmd/quickstart is a 13-phase ~260-line end-to-end
walk-through. The two pieces drift independently — the README
was using outdated ceremony (panic instead of log.Fatal,
IIFE-wrapped defer Close, a 5-second WithTimeout that could
never fire in a happy-path run, a // Output: "..." comment
that's a godoc convention that doesn't apply to cmd/ binaries)
while the program had been refined to cover every exported
symbol in a single go run.

Split cmd/quickstart into two binaries so each piece has a
single, clear job:

  cmd/quickstart          (NEW 73-line minimum viable example)
    Mirrors README's Quick Start 1:1. Single broker, single
    publisher, single subscriber, single topic, one Publish,
    one Receive. Synchronous Publish means the 200ms
    WithTimeout demonstrably resets on the happy path
    (select hits the Ch branch, never the ErrCh one).
    `go run ./cmd/quickstart` prints:

        Received: CPU over 90%!

  cmd/quickstart-e2e      (renamed, content unchanged)
    The 13-phase end-to-end walk-through that previously
    lived at cmd/quickstart/main.go. Same surface, same
    phase numbering, same output. Only the trailing
    "quickstart: ok" was renamed to "quickstart-e2e: ok"
    to match the new binary name.

README's Quick Start is rewritten to be 1:1 with the new
simplified binary. It now points at cmd/quickstart-e2e for
the long-form walk-through instead of being a single
self-contained snippet.

cmd/quickstart/README.md is rewritten to describe the
minimal binary and point at cmd/quickstart-e2e/README.md
for the longer walk-through. cmd/quickstart-e2e/README.md
is the previous contents of cmd/quickstart/README.md
with internal references updated to the new path.

No library code changes; existing unit tests in pubsub/
remain the contract.

Verified:
  go run ./cmd/quickstart      Received: CPU over 90%!
  go run ./cmd/quickstart-e2e  quickstart-e2e: ok
  go build ./...               ok
  go vet ./...                 clean
  gofmt -l $(git ls-files)    clean
  go test -race -count=1 ./... 3.1s, green
